+++ api/tests/src/api/applications/test_application_routes.py
@@ -16,10 +16,7 @@ from src.constants.lookup_constants import (
     Privilege,
 )
 from src.db.models.competition_models import Application, ApplicationForm, ApplicationStatus
-from src.db.models.competition_models import Form as FormModel
 from src.db.models.user_models import ApplicationUser
-from src.form_schema.forms import init_form_registry
-from src.form_schema.registry.form_template_registry import FormTemplateKey, form_template_registry
 from src.validation.validation_constants import ValidationErrorType
 from tests.lib.application_test_utils import create_user_in_app
 from tests.lib.organization_test_utils import create_user_in_org
@@ -58,40 +55,6 @@ SIMPLE_ATTACHMENT_JSON_SCHEMA = {
 }
 
 SIMPLE_ATTACHMENT_RULE_SCHEMA = {"attachment_field": {"gg_validation": {"rule": "attachment"}}}
-
-
-@pytest.fixture
-def create_test_form(db_session):
-    """Factory fixture for custom-schema test forms with automatic registry cleanup.
-    # TODO(#10274): remove db_session.add + flush once the form table is dropped
-    """
-    init_form_registry()
-
-    registered_key = None
-
-    def _make(**kwargs) -> FormModel:
-        nonlocal registered_key
-        form = FormModel(
-            form_id=uuid.uuid4(),
-            form_name=kwargs.get("form_name", "Test Form"),
-            short_form_name=kwargs.get("short_form_name", "TestForm"),
-            form_version="1.0",
-            agency_code="SGG",
-            form_json_schema=kwargs.get("form_json_schema", {"type": "object", "properties": {}}),
-            form_ui_schema={},
-            form_rule_schema=kwargs.get("form_rule_schema", None),
-            json_to_xml_schema=kwargs.get("json_to_xml_schema", None),
-        )
-        db_session.add(form)
-        db_session.flush()
-        form_template_registry.register(form, major_version=1)
-        registered_key = FormTemplateKey(form.form_id, 1)
-        return form
-
-    yield _make
-
-    if registered_key:
-        form_template_registry._registry.pop(registered_key, None)
 
 
 def test_application_start_success(
