+++ api/tests/src/services/applications/test_add_organization_to_application.py
@@ -9,9 +9,6 @@ from src.constants.lookup_constants import (
     CompetitionOpenToApplicant,
     Privilege,
 )
-from src.db.models.competition_models import Form as FormModel
-from src.form_schema.forms import init_form_registry
-from src.form_schema.registry.form_template_registry import FormTemplateKey, form_template_registry
 from src.services.applications.add_organization_to_application import (
     add_organization_to_application,
 )
@@ -28,40 +25,6 @@ from tests.src.db.models.factories import (
     RoleFactory,
     UserFactory,
 )
-
-
-@pytest.fixture
-def create_test_form(db_session):
-    """Factory fixture for custom-schema test forms with automatic registry cleanup.
-    # TODO(#10274): remove db_session.add + flush once the form table is dropped
-    """
-    init_form_registry()
-
-    registered_key = None
-
-    def _make(**kwargs) -> FormModel:
-        nonlocal registered_key
-        form = FormModel(
-            form_id=uuid.uuid4(),
-            form_name=kwargs.get("form_name", "Test Form"),
-            short_form_name=kwargs.get("short_form_name", "TestForm"),
-            form_version=kwargs.get("form_version", "1.0"),
-            agency_code="SGG",
-            form_json_schema=kwargs.get("form_json_schema", {"type": "object", "properties": {}}),
-            form_ui_schema={},
-            form_rule_schema=kwargs.get("form_rule_schema", None),
-            json_to_xml_schema=kwargs.get("json_to_xml_schema", None),
-        )
-        db_session.add(form)
-        db_session.flush()
-        form_template_registry.register(form, major_version=1)
-        registered_key = FormTemplateKey(form.form_id, 1)
-        return form
-
-    yield _make
-
-    if registered_key:
-        form_template_registry._registry.pop(registered_key, None)
 
 
 def test_add_organization_to_application_success(
