+++ api/tests/src/services/xml_generation/test_submission_xml_assembler.py
@@ -1,16 +1,12 @@
 """Tests for submission XML assembler."""
 
-import uuid
 from datetime import date
 
 import grants_shared.adapters.db as db
 import pytest
 from lxml import etree as lxml_etree
 
-from src.db.models.competition_models import Form as FormModel
-from src.form_schema.forms import init_form_registry
 from src.form_schema.forms.sf424 import FORM_XML_TRANSFORM_RULES
-from src.form_schema.registry.form_template_registry import FormTemplateKey, form_template_registry
 from src.services.xml_generation.constants import Namespace
 from src.services.xml_generation.submission_xml_assembler import SubmissionXMLAssembler
 from tests.src.db.models.factories import (
@@ -23,40 +19,6 @@ from tests.src.db.models.factories import (
     OpportunityAssistanceListingFactory,
     OpportunityFactory,
 )
-
-
-@pytest.fixture
-def create_test_form(db_session):
-    """Factory fixture for custom-schema test forms with automatic registry cleanup.
-    # TODO(#10274): remove db_session.add + flush once the form table is dropped
-    """
-    init_form_registry()
-
-    registered_key = None
-
-    def _make(**kwargs) -> FormModel:
-        nonlocal registered_key
-        form = FormModel(
-            form_id=uuid.uuid4(),
-            form_name=kwargs.get("form_name", "Test Form"),
-            short_form_name=kwargs.get("short_form_name", "TestForm"),
-            form_version=kwargs.get("form_version", "1.0"),
-            agency_code="SGG",
-            form_json_schema=kwargs.get("form_json_schema", {"type": "object", "properties": {}}),
-            form_ui_schema={},
-            form_rule_schema=kwargs.get("form_rule_schema", None),
-            json_to_xml_schema=kwargs.get("json_to_xml_schema", None),
-        )
-        db_session.add(form)
-        db_session.flush()
-        form_template_registry.register(form, major_version=1)
-        registered_key = FormTemplateKey(form.form_id, 1)
-        return form
-
-    yield _make
-
-    if registered_key:
-        form_template_registry._registry.pop(registered_key, None)
 
 
 class TestSubmissionXMLAssembler:
